Fix import paths for deeply-imported packages
Use the task's package's import path, since the import path of a task
may be different than the path of the package we imported it from.  This
is the case for a->b->c, since a imports task c via b, but the import
path should be that of c.

Add test coverage