Revise required-overlays check
It is inappropriate to check for required-overlays at parse-time,
because we may not be interested in the task with the required-overlays.

Specifically, a case arose where a build task required linux/macosx
overlays, but a clean task did not.  However, the user was required to
run clean with the platform overlay for the (unused) build task to be
valid.

The resolution here is to move the required overlays check to a separate
API.  The downside here is that callers have to opt-in to this behavior.

An alternate design would be to intercept this in the task subscript,
but that was considered too invasive.

Resolves https://github.com/AnarchyTools/atbuild/issues/66