Convert Task.allKeys to computed property
Currently, allKeys can get out of sync with the actual task.  This can occur if you use the subscript setter.

This does not arise very often, but can occur if you have an attool (which has no fixed set of keys, so it uses allKeys) *and* you are setting keys from overlays or via some other method that modifies the keys for tasks after their creation.